+++ src/workspace/WorkSpaceMain.jsx
@@ -1,19 +1,10 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import { Header, Body, Terminal } from './components';
-
+import { Header, Body } from './components';
 const useStyles = makeStyles((theme) => ({
   root: {
     height: '100vh',
     overflow: 'hidden',
-  },
-  body: {
-    height: '71vh',
-    borderBottom: '1px solid gray',
-    position: 'relative',
-  },
-  terminal: {
-    height: '25vh',
   },
 }));
 
@@ -23,12 +14,7 @@ function WorkSpaceMain(props) {
   return (
     <div className={classes.root}>
       <Header />
-      <div className={classes.body}>
-        <Body />
-      </div>
-      <div className={classes.terminal}>
-        <Terminal />
-      </div>
+      <Body />
     </div>
   );
 }
