+++ src/workspace/components/BottomPanel.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Terminal } from './';
+
+function BottomPanel(props) {
+  return <Terminal />;
+}
+
+export default BottomPanel;
