+++ src/workspace/components/Editor.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-function CodeEditorView(props) {
+function Editor(props) {
   const { fileName } = props;
 
   return <div>{fileName}</div>;
 }
 
-export default CodeEditorView;
+export default Editor;
