+++ src/workspace/components/FileTree.jsx
@@ -1,6 +1,49 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import { FileTreeView, FileTab } from './';
+import TreeView from '@material-ui/lab/TreeView';
+import TreeItem from '@material-ui/lab/TreeItem';
+
+import { useDispatch } from 'react-redux';
+import { selectFile } from '../modules/reducers';
+
+const data = {
+  id: 'root',
+  name: 'Project',
+  children: [
+    {
+      id: '1',
+      name: 'public',
+      type: 'directory',
+    },
+    {
+      id: '3',
+      name: 'src',
+      type: 'directory',
+      children: [
+        {
+          id: '4',
+          name: 'App.js',
+          type: 'file',
+        },
+        {
+          id: '5',
+          name: 'index.js',
+          type: 'file',
+        },
+        {
+          id: '6',
+          name: 'App.css',
+          type: 'file',
+        },
+        {
+          id: '7',
+          name: 'App.test.js',
+          type: 'file',
+        },
+      ],
+    },
+  ],
+};
 
 const useStyles = makeStyles({
   root: {
@@ -11,11 +54,38 @@ const useStyles = makeStyles({
 function FileTree(props) {
   const classes = useStyles();
 
+  // redux로 codeEditor에 보여질 파일관리
+  const dispatch = useDispatch();
+  const onSelectFile = (name) => dispatch(selectFile(name));
+
+  const handleClick = (e) => {
+    if (e.type === 'file') {
+      onSelectFile(e.name);
+    }
+  };
+
+  const renderTree = (nodes) => (
+    <TreeItem
+      key={nodes.id}
+      nodeId={nodes.id}
+      label={nodes.name}
+      onClick={() => handleClick(nodes)}
+    >
+      {Array.isArray(nodes.children)
+        ? nodes.children.map((node) => renderTree(node))
+        : null}
+    </TreeItem>
+  );
+
   return (
-    <div className={classes.root}>
-      <FileTab />
-      <FileTreeView />
-    </div>
+    <TreeView
+      className={classes.root}
+      defaultCollapseIcon={'>'}
+      defaultExpanded={['root']}
+      defaultExpandIcon={'+'}
+    >
+      {renderTree(data)}
+    </TreeView>
   );
 }
