+++ src/workspace/components/Header.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 
-import { HeaderMenuList } from './';
+import { MenuBar } from './';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -22,7 +22,7 @@ function Header(props) {
     <Grid container className={classes.root}>
       <div className={classes.logo}>울림 IDE</div>
       <Grid item xs={11}>
-        <HeaderMenuList />
+        <MenuBar />
       </Grid>
     </Grid>
   );
