+++ src/workspace/components/index.js
@@ -1,14 +1,18 @@
 export { default as Header } from './Header';
-export { default as HeaderMenuList } from './HeaderMenuList';
-export { default as HeaderMenuBtn } from './HeaderMenuBtn';
+export { default as MenuBar } from './MenuBar';
+export { default as MenuItem } from './MenuItem';
+
 export { default as Body } from './Body';
+
+export { default as Workspace } from './Workspace';
+export { default as OpenFileTab } from './OpenFileTab';
+export { default as Editor } from './Editor';
+export { default as QuickIconTab } from './QuickIconTab';
+
+export { default as Explorer } from './Explorer';
+export { default as FileTreeTab } from './FileTreeTab';
 export { default as FileTree } from './FileTree';
-export { default as FileTab } from './FileTab';
-export { default as FileTreeView } from './FileTreeView';
+export { default as CommandList } from './CommandList';
 
+export { default as BottomPanel } from './BottomPanel';
 export { default as Terminal } from './Terminal';
-export { default as TerminalBody } from './TerminalBody';
-export { default as Command } from './Command';
-
-export { default as CodeEditorView } from './CodeEditorView';
-export { default as CodeEditor } from './CodeEditor';
