--- src/workspace/containers/index.js
@@ -1,1 +1,0 @@
-export { default as CurrentFileContainer } from './CurrentFileContainer';
